messages_app/requirements.txt
amqp==5.2.0
anyio==4.3.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
autopep8==2.0.4
billiard==4.2.0
celery==5.3.6
certifi==2024.2.2
channels==4.0.0
channels-redis==4.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.3
Django==4.2.10
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-cors-headers==4.3.1
django-silk==5.1.0
django-timezone-field==6.1.0
djangorestframework==3.14.0
drf-spectacular==0.27.1
drf-spectacular-sidecar==2024.2.1
ecdsa==0.19.0
exceptiongroup==1.2.0
gprof2dot==2022.7.29
h11==0.14.0
httptools==0.6.1
idna==3.6
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kombu==5.3.5
msgpack==1.0.8
packaging==23.2
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
pyasn1==0.6.0
pycodestyle==2.11.1
PyJWT==2.8.0
python-crontab==3.0.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python-jose==3.3.0
pytz==2024.1
PyYAML==6.0.1
redis==5.0.2
referencing==0.33.0
requests==2.31.0
rpds-py==0.18.0
rsa==4.9
shortuuid==1.0.12
six==1.16.0
sniffio==1.3.1
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.9.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
uvicorn==0.27.1
uvloop==0.19.0
vine==5.1.0
watchfiles==0.21.0
wcwidth==0.2.13
websockets==12.0
whitenoise==6.6.0
